refactor: update HeaderTopBar component
