+++ src/Layout/Header/modules/HeaderTopBar/HeaderTopBar.module.scss
@@ -1,0 +1,14 @@
+@import '@/styles';
+
+.top {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+
+  padding: 24px 0;
+  padding-right: 77px;
+
+  .logo {
+    font-size: 0;
+  }
+}
